Reestructuracion de archivos del proyecto
