--- index.py
@@ -1,51 +1,0 @@
-""" index file for REST APIs using Flask """
-import os
-import sys
-import requests
-from flask import jsonify, request, make_response, send_from_directory
-
-ROOT_PATH = os.path.dirname(os.path.realpath(__file__))
-os.environ.update({'ROOT_PATH': ROOT_PATH})
-sys.path.append(os.path.join(ROOT_PATH, 'server'))
-
-import logger
-from app import app
-from app import users
-
-# Create a logger object to log the info and debug
-LOG = logger.get_root_logger(os.environ.get(
-    'ROOT_LOGGER', 'root'), filename=os.path.join(ROOT_PATH, 'output.log'))
-
-# Port variable to run the server on.
-PORT = os.environ.get('PORT')
-
-
-@app.errorhandler(404)
-def not_found(error):
-    """ error handler """
-    LOG.error(error)
-    # return make_response(jsonify({'error': 'Not found'}), 404)
-    return send_from_directory('template', '404.html')
-
-
-@app.route('/')
-# Esta funcion se llama index2 porque si se llamara 
-# index rompe el codigo por dos funciones del mismo nombre
-def index2():
-    """ static files serve """
-    return send_from_directory('template', 'index.html')
-
-
-@app.route('/<path:path>')
-def static_proxy(path):
-    """ static folder serve """
-    file_name = path.split('/')[-1]
-    dir_name = os.path.join('template', '/'.join(path.split('/')[:-1]))
-    return send_from_directory(dir_name, file_name)
-
-
-if __name__ == '__main__':
-    LOG.info('running environment: %s', os.environ.get('ENV'))
-    # Debug mode if development env
-    app.config['DEBUG'] = os.environ.get('ENV') == 'development'
-    app.run(host='0.0.0.0', port=int(PORT))  # Run the app
