--- server/app/__init__.py
@@ -1,28 +1,0 @@
-import os
-import datetime
-import json
-
-from bson.objectid import ObjectId
-from flask import Flask
-from flask_pymongo import PyMongo
-
-
-class JSONEncoder(json.JSONEncoder):
-    # extend json encoder for compatibility with mongoDB
-    def default(self, o):  # pylint: disable=E0202
-        if isinstance(o, ObjectId):
-            return str(o)
-        if isinstance(o, datetime.datetime):
-            return str(o)
-        return json.JSONEncoder.default(self, o)
-
-
-# create the flask object
-app = Flask(__name__)  # pylint: disable=C0103
-""" add mongo url to flask config, so that
-flask_pymongo can use it to make connection"""
-app.config['MONGO_URI'] = os.environ.get('DB')
-mongo = PyMongo(app)  # pylint: disable=C0103
-""" use the modified encoder class to handle ObjectId
-& datetime object while jsonifying the response """
-app.json_encoder = JSONEncoder
