--- server/logger/__init__.py
@@ -1,2 +1,0 @@
-# flake8: noqa
-from .logger import get_root_logger, get_child_logger
